Use all sykmeldinger in graf regardless of virksomhet
This also affects UtdragFraSykefravaret, since that also uses the sykmeldingerInnenforOppfolgingstilfelle method.
Now it's up to the user of the util method to decide the limitations on which sykmeldinger should count, since it only checks dates as the method name implies.
Also add test, it didn't actually verify the filtering by virksomhet.