IS-1796: Add support for new question in soknad
New question "TIL_SLUTT" replaces "VAER_KLAR_OVER_AT" and
"BEKREFT_OPPLYSNINGER".
"bekreft" is now an underspørsmål in "TIL_SLUTT".
The old format is still supported so legacy søknader will look the same
as before.
We assume the new svartype "bekreftelsespunkter" is a list of "svar"
that needs to be displayed as a list of html strings (to format urls
correctly).